+++ src/uses/fetchTransactions.js
@@ -1,0 +1,17 @@
+import { ref } from "vue";
+
+export default function () {
+    const transactions = ref([]);
+    const error = ref(null);
+    const fetchData = async () => {
+        try {
+            const response = await fetch("http://localhost:3000/transactions")
+            if (!response.ok) throw new Error("Something went wrong.");
+            transactions.value = await response.json();
+        } catch (e) {
+            error.value = e;
+        }
+    };
+
+    return { transactions, error, fetchData };
+}
